+++ api/app/controllers/api/v1/user/users_controller.rb
@@ -1,0 +1,26 @@
+module Api
+  module V1
+    module User
+      class UsersController < ApplicationController
+
+        def index
+          @user = User.all
+          render json: @user
+        end
+
+        def update
+          if @user.update(user_params)
+            render json: @user
+          else
+            render json: @user.errors
+          end
+        end
+
+        private
+          def user_params
+            params.require(:user).permit(:name, :email, :image, :height, :sex, :birthday, :training_type)
+          end
+      end
+    end
+  end
+end
