--- api/app/controllers/users_controller.rb
@@ -1,28 +1,0 @@
-class UsersController < ApplicationController
-  before_action :authenticate_user!
-
-  def index
-    @user = User.all
-    render json: @user
-  end
-
-  def update
-    @user = User.find(params[:id])
-    @user.update(user_params)
-    render json: @user
-  end
-
-  def destroy
-    @post = Post.find(params[:id])
-    if @post.destroy
-      head :no_content, status: :ok
-    else
-      render json: @post.errors, status: :unprocessable_entity
-    end
-  end
-
-  private
-    def user_params
-      params.require(:user).permit(:name, :email, :image, :height, :sex, :birthday, :training_type)
-    end
-end
